Created a new abstract class Storage which StorageFile depends on.
